# docs/swagger.yaml
definitions:
  main.Transaction:
    properties:
      amount:
        type: number
      category:
        description: категория - еда, одежда и тд
        type: string
      createdAt:
        description: автоматически создается GORM
        type: string
      date:
        type: string
      description:
        description: может быть пустым
        type: string
      id:
        type: integer
      type:
        description: тип транзакции - трата или расход
        type: string
      userID:
        description: Привязка к пользователю
        type: integer
    type: object
  main.authRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: API for tracking personal finance transactions
  title: Finance tracker API
paths:
  /api/balance:
    get:
      consumes:
      - application/json
      description: Calculate and return the balance for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Balance response
          schema:
            additionalProperties:
              type: number
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user balance
      tags:
      - transactions
  /api/transactions:
    get:
      consumes:
      - application/json
      description: Retrieve all transactions for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            items:
              $ref: '#/definitions/main.Transaction'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all transactions
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Create a new transaction for the authenticated user
      parameters:
      - description: Transaction data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/main.Transaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created transaction
          schema:
            $ref: '#/definitions/main.Transaction'
        "400":
          description: Invalid request body or parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new transaction
      tags:
      - transactions
  /api/transactions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a transaction by ID for the authenticated user
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a transaction
      tags:
      - transactions
    put:
      consumes:
      - application/json
      description: Fully update a transaction by ID for the authenticated user
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Full transaction data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/main.Transaction'
      produces:
      - application/json
      responses:
        "200":
          description: Updated transaction
          schema:
            $ref: '#/definitions/main.Transaction'
        "400":
          description: Invalid request body or parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Transaction not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a transaction
      tags:
      - transactions
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/main.authRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Token response
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid email or password
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user with email and password
      parameters:
      - description: User credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.authRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success response
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or email already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
swagger: "2.0"